davinciflow/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
asyncpg==0.30.0
aiosqlite==0.21.0
alembic==1.15.2
pydantic==2.11.1
pydantic-settings==2.8.1
httpx==0.28.1
pandas==2.2.3
pyarrow==19.0.1
confluent-kafka==2.8.2
paho-mqtt==2.1.0
minio==7.2.15
boto3==1.37.12
structlog==25.2.0
python-multipart==0.0.20
websockets==15.0.1
greenlet==3.1.1

